requirements.txt
certifi==2018.4.16
chardet==3.0.4
click==6.7
docker==3.4.1
docker-pycreds==0.3.0
idna==2.7
Jinja2==2.10
jsonschema==2.6.0
MarkupSafe==1.0
packaging==17.1
pluggy==0.7.1
py==1.5.4
pyparsing==2.2.0
PyYAML==3.13
requests==2.19.1
six==1.11.0
tox==3.1.3
urllib3==1.23
virtualenv==16.0.0
websocket-client==0.48.0
yapf==0.22.0
